requirements.txt
asgiref==3.11.0
black==26.1.0
click==8.3.1
Django==6.0.1
django-cors-headers==4.4.0
django-otp==1.7.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.7
isort==7.0.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pyotp==2.9.0
pytokens==0.4.1
qrcode==8.2
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
